Finish refacto user api change serial to bigint

+++ schema.sql
@@ -38,7 +38,7 @@ VALUES (0
 
 CREATE TABLE users
 (
-    id              SERIAL NOT NULL UNIQUE,
+    id              BIGINT NOT NULL UNIQUE,
     name            VARCHAR(256) NOT NULL,
     surname         VARCHAR(256) NOT NULL,
     email           VARCHAR(256) NOT NULL UNIQUE,
